Move json mcdoc checker from JE to json package
