+++ packages/json/src/index.ts
@@ -6,6 +6,7 @@ import * as completer from './completer/index.js'
 import * as formatter from './formatter/index.js'
 import * as parser from './parser/index.js'
 
+export * as checker from './checker/index.js'
 export * as colorizer from './colorizer/index.js'
 export * as completer from './completer/index.js'
 export * as formatter from './formatter/index.js'
