+++ packages/mcdoc-cli/tsconfig.json
@@ -11,5 +11,8 @@
 		{
 			"path": "src"
 		},
+		{
+			"path": "test"
+		}
 	]
 }
